Evaluation (#53)
* feat: mobility evaluation

* perf: skip one piece to make color() and piece_type() faster

* refactor: some comments and Magic Number cleanup

* perf(position): optimize check queries for the active side

- Rename `is_in_check(Side)` to `is_side_in_check(Side)` for general use
- Add parameterless `is_in_check(&self) -> bool` returning
`self.state.in_check`
  directly for the side-to-move, removing redundant check evaluations
  and function call overhead on hot paths
- Inline hot-path lookup methods in `KillerMoves` and `HistoryMoves`
- Add `Position` and `StateInfo` size-of measurements to
`test_alignment`

* feat: defender bonus

* chore: prepare for texel tuning

Introduces a tunable evaluation parameter system to support automated
tuning
tools (e.g. SPSA) by parameterizing weights rather than using hardcoded
constants.

- Create `EvalParams` struct in `src/eval/eval_params.rs` packaging
material,
  piece-square tables, mobility, and defender bonuses
- Add `DEFAULT_PARAMS` containing original weight configurations
- Implement `to_vector` and `update_from_vector` for parameter
serialization
- Expose `evaluate_with_params` and helper functions to compute
material,
  mobility, and defender scores based on custom parameter configurations
- Make static PST and mobility tables visible to `crate::eval`

* perf(tuner): multithread MSE calculation and add progress reporting

- Parallelize MSE computation using `std::thread::scope` to chunk
positions
  across worker threads matching the available CPU thread count
- Implement inline console progress reporting with ETA estimation,
elapsed
  time, and MSE indicators updated every 100 parameters

* feat(tuner): add Ctrl+C signal handler for graceful shutdown

- Integrate the `ctrlc` crate to capture interrupt signals during tuning
- Set up an atomic `running` flag checked at the start of each iteration
  and parameter search step to abort processing cleanly on Ctrl+C
- Save and print the best-known parameters upon graceful termination

* fix: update tuner to print good results

* refactor: use a seperate mechanism to store piece material value and update params with tuner output

* fix: make eval params use eval params as default

* refactor(eval): clean up and reorganize evaluation functions

- Split material value lookup tables and logic to
`piece_material_value.rs`
- Split defender bonus tables and logic to `defender_bonus.rs`
- Add `Piece::SEPARATOR` constant and `Position::piece_type_count`
helper
  to count pieces of both colors for a given type
- Simplify `calculate_phase` in `eval/mod.rs` by utilizing the new
  `piece_type_count` helper
- Migrate the game phase type representation from `i32` to `u8`
- Delegate `Position` evaluation and phase methods to their
corresponding
  `eval` modules to decouple the core structs from evaluation details

* chore(eval): update evaluation tables with tuned weights

Updates default evaluation weights in the static PSTs, mobility, and
defender count tables with optimized values generated by the tuner.

- Update `PIECE_SQUARE_TABLE_*_TAPERED` tables in
`piece_square_tables.rs`
- Update `*_MOBILITY_BONUS` tables in `mobility_tables.rs`
- Update `ADVISOR_COUNT_BONUS` and `BISHOP_COUNT_BONUS` in
`defender_bonus.rs`
- Clean up output formatting in `tuner.rs` for generating copy-pasteable
blocks

* fix: make tuner stop on convergence and do symmetric tuning for pst

* feat: make tuner uses Adam

* chore: make sane defaults for tuning

* chore: add update_eval.py script and update values from texel tuning

* chore: format files

* chore: update script to download latest Sylvan binary

* feat: add game generation script for texel

* refactor: use pre-existing function call

* refactor: remove thiserror dependency

* refactor: use Result<> instead of Option<> for command parsing

* refactor: remove unnecessary binary

* refactor: better visibility

* refactor: better visibility #2

* refactor: rename zobrist_hash to hash

* refactor: add indexing support for enums

* refactor: remove strum dependency

* refactor: use defined utils function

* refactor(core): implement type-safe array indexing for core types

Eliminates verbose `as usize` casts across the codebase by implementing
`Index`
and `IndexMut` directly for core enum types (`PieceType`, `Piece`,
`Side`, `Square`).

- Add `impl_index!` macro in `src/core/types/mod.rs` and apply it to
core types
- Remove manual `Index` implementations for `PieceType`, `Piece`,
`Side`, and `Square`
- Replace direct array index casts with type-safe enum indexing across
position logic,
  move generation, and search history
- Use `Square::mirrored()` in `piece_square_tables.rs` instead of
repeating
  manual rank/file vertical and horizontal mirroring arithmetic
- Simplify `compute_defender_bonus` to directly sum and return the
bonuses
- Mark query helpers in `queries.rs` as `const fn`

* style: cargo fmt

+++ src/search/history_moves.rs
@@ -1,5 +1,3 @@
-use strum::EnumCount;
-
 use crate::core::{Move, Side, Square};
 
 const DECAY_RATE: i32 = 8;
@@ -19,20 +17,23 @@ impl HistoryMoves {
         }
     }
 
+    #[inline]
     pub fn get(&self, color: Side, mv: Move) -> i32 {
-        self.table[color as usize][mv.from() as usize][mv.to() as usize]
+        self.table[color][mv.from()][mv.to()]
     }
 
+    #[inline]
     pub fn increase(&mut self, color: Side, mv: Move, depth: i8) {
         let depth = depth as i32;
         let bonus = depth * depth;
-        self.table[color as usize][mv.from() as usize][mv.to() as usize] += bonus;
+        self.table[color][mv.from()][mv.to()] += bonus;
     }
 
+    #[inline]
     pub fn decrease(&mut self, color: Side, mv: Move, depth: i8) {
         let depth = depth as i32;
         let bonus = depth * depth;
-        self.table[color as usize][mv.from() as usize][mv.to() as usize] -= bonus;
+        self.table[color][mv.from()][mv.to()] -= bonus;
     }
 
     pub fn decay(&mut self) {
